Mainpage.tsx 에 file drag and drop 설정
