+++ src/components/Chat/ChatInput.tsx
@@ -5,7 +5,6 @@ import { RootState } from "../../app/store";
 import { at, WsUrl_chat } from "../../variable/cookie";
 import { UpdateChat } from "../../variable/UpdateChatContextSlice";
 import { findUserDataInClickedChannel } from "../../variable/ClickedChannelSlice";
-import DragDrop from "./FileDragDrop";
 
 function ChatInput(props: any) {
   const [msg, setmsg] = useState("");
@@ -69,7 +68,6 @@ function ChatInput(props: any) {
   return (
     <ChatInputContainer>
       <form>
-        <DragDrop />
         <input ref={inputRef} onChange={e => setmsg(e.target.value)} placeholder={`Message #`} />
         <button hidden type="submit" onClick={sendMessage}>
           SEND
