--- src/components/Chat/drag_drop.css
@@ -1,85 +1,0 @@
-  /* 드래그 앤 드롭 */
-  .div_images {
-      padding: 10px;
-      font-size: 0.8em;
-      color: grey;
-      margin: 10px 0 10px 30px;
-      box-shadow: 0 0 5px #cccccc;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-  }
-
-  .div_preview {
-      display: flex;
-      flex-direction: row;
-      justify-content: center;
-      padding: 10px;
-  }
-
-  /* 이미지 프리뷰 */
-  .div_preview>img {
-      height: 100px;
-      width: 100px;
-      cursor: pointer;
-  }
-
-  /* pdf 이미지 클릭 */
-  .div_preview>div {
-      cursor: pointer;
-  }
-
-  /* pdf 이미지 프리뷰 */
-  .div_preview>div>embed {
-      height: 100px;
-      width: 100px;
-      display: block;
-      z-index: -1;
-      position: relative;
-  }
-
-  /* 이미지 삭제 버튼 */
-  .div_preview+button {
-      border: none;
-      color: white;
-      background-color: rgb(223, 71, 71);
-      cursor: pointer;
-      margin-top: 10px;
-  }
-
-  .div_preview+button:hover {
-      animation: outline 1s ease;
-  }
-
-  #input_file {
-      display: none;
-  }
-
-  /* 이미지 업로드 버튼 */
-  #input_file+label {
-      border: 1px solid #cccccc;
-      border-radius: 12px;
-      padding: 0 10px 0 10px;
-      cursor: pointer;
-  }
-
-  #input_file+label:hover {
-      background-color: #cccccc;
-      color: white;
-  }
-
-  @keyframes outline {
-      from {
-          outline: #e4e4e4 solid;
-          outline-offset: 10px;
-          transform: scale(1)
-      }
-
-      to {
-          outline: transparent solid;
-          outline-offset: 0px;
-          transform: scale(1.05);
-          box-shadow: 0 2px 10px #e4e4e4;
-      }
-  }
