+++ src/pages/Mainpage.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { ChangeEvent, DragEvent, useCallback, useEffect, useState } from "react";
 import Header from "../components/Header";
 import Sidebar from "../components/Sidebar/Sidebar";
 import Chat from "../components/Chat/Chat";
@@ -15,7 +15,6 @@ import { getMyProfile } from "../variable/MyProfileSlice";
 import { SelectWorkspace } from "../components/Workspace/Workspace";
 import PlusModal from "../components/Workspace/PlusModal";
 import ChannelSetting from "../components/Channel/ChannelSetting";
-import DragDrop from "../components/Chat/FileDragDrop";
 
 const Mainpage = () => {
   const dispatch = useDispatch();
@@ -68,10 +67,69 @@ const Mainpage = () => {
   const onClickToggleModal = useCallback(() => {
     setOpenModal(!isOpenModal);
   }, [isOpenModal]);
+
+  const [imageList, setImageList] = useState<Array<File>>([]);
+
+  // 이미지 파일 처리 input
+  const onInputFile = (e: ChangeEvent<HTMLInputElement>) => {
+    e.preventDefault();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    handleFiles(e.target.files);
+  };
+
+  // 파일 처리 ondrop
+  const onDropFiles = (e: DragEvent<HTMLDivElement>) => {
+    console.log({ e }, e.dataTransfer.files);
+    e.preventDefault();
+
+    handleFiles(e.dataTransfer.files);
+  };
+
+  const handleFiles = async (files: FileList) => {
+    let fileList: Array<File> = [];
+    for (let i = 0; i < files.length; i++) {
+      const file: File = files[i];
+      const format: string = `${file.name.split(".").slice(-1)}`.toUpperCase();
+      if (format === "JPG" || format === "JPEG" || format === "PNG" || format === "PDF") {
+        console.log(file);
+        if ((await AtVerify()) == 200) {
+          fileList = [...fileList, file];
+
+          await axios.post(
+            `${backUrl}file/`,
+            {
+              file: file,
+            },
+            {
+              headers: {
+                Authorization: `Bearer ${at}`,
+              },
+            },
+          );
+        } else {
+          alert(`지원하지 않는 포맷입니다: ${file.name} / FORMAT ${format}`);
+          return;
+        }
+      } else {
+        alert(`지원하지 않는 포맷입니다: ${file.name} / FORMAT ${format}`);
+        return;
+      }
+    }
+    if (fileList.length > 0) {
+      setImageList(fileList);
+    }
+  };
+
+  // 없으면 drop 작동안됨
+  const dragOver = (e: DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+  };
+
   return (
     <>
       <Logout />
-      <AppBody>
+      <AppBody onDrop={onDropFiles} onDragOver={dragOver}>
         <Header />
         <SelectWorkspaces>
           {Workspace.map((element, i) => {
